Fix a serious bug related to retain lines
Before this change the retain lines were generated from rsnapshot::interval
definitions. While this was cool and convenient, it had the side-effect of
ordering the retain lines in rsnapshot.conf somewhat randomly. As rsnapshot
_depends_ on correct ordering of these lines, but does not fail if they're not
in order, this could cause rsnapshot to rotate it's own backups into oblivion.
Switched to defining the retain lines as an array in the main rsnapshot class
instead; this approach will produce consistent ordering every time. Also renamed
rsnapshot::interval as rsnapshot::cron to make this change more obvious.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>