Add class for updating backup marker files automatically
The idea is that cron updates a marker file in a directory that is being backed
up. The updates happen regularly, by default daily, before backups are taken. On
the rsnapshot server side the timestamp on the marker files allows gathering
metrics about backup age, e.g. with Prometheus Node Exporter's Textfile
Collector. If marker files are too old we can deduce with high probability that
backups are failing and can send an alert.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>